BackEnd/requirement.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
astunparse==1.6.3
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
distro==1.9.0
Django==5.2.9
Flask==3.1.2
flatbuffers==25.9.23
fonttools==4.58.0
gast==0.7.0
google-auth==2.43.0
google-genai==1.55.0
google-pasta==0.2.0
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
httpcore==1.0.9
httpx==0.28.1
idna==3.11
image==1.5.33
itsdangerous==2.2.0
Jinja2==3.1.6
keras==3.12.0
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.3
mdurl==0.1.2
ml_dtypes==0.5.4
namex==0.1.0
numpy==2.2.6
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
protobuf==6.33.1
PuLP==3.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pydataset==0.2.0
Pygments==2.19.2
pyparsing==3.2.3
pypi==2.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
rich==14.2.0
rsa==4.9.1
scipy==1.16.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.4
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
tensorflow_keras==0.1
termcolor==3.2.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.1.4
wrapt==2.0.1
